+++ vkcs/dc/resource_vrrp_address.go
@@ -13,10 +13,15 @@ import (
 	"github.com/vk-cs/terraform-provider-vkcs/vkcs/internal/services/dc/v2/vrrpaddresses"
 	"github.com/vk-cs/terraform-provider-vkcs/vkcs/internal/services/networking"
 	"github.com/vk-cs/terraform-provider-vkcs/vkcs/internal/util"
+	"github.com/vk-cs/terraform-provider-vkcs/vkcs/internal/util/errutil"
 )
 
 // Ensure the implementation satisfies the desired interfaces.
-var _ resource.Resource = &VRRPAddressResource{}
+var (
+	_ resource.Resource                = &VRRPAddressResource{}
+	_ resource.ResourceWithConfigure   = &VRRPAddressResource{}
+	_ resource.ResourceWithImportState = &VRRPAddressResource{}
+)
 
 func NewVRRPAddressResource() resource.Resource {
 	return &VRRPAddressResource{}
@@ -281,7 +286,9 @@ func (r *VRRPAddressResource) Delete(ctx context.Context, req resource.DeleteReq
 	id := data.ID.ValueString()
 
 	err = vrrpaddresses.Delete(networkingClient, id).ExtractErr()
-	if err != nil {
+	if errutil.IsNotFound(err) {
+		return
+	} else if err != nil {
 		resp.Diagnostics.AddError("Unable to delete resource vkcs_dc_vrrp_address", err.Error())
 		return
 	}